refactor(Heading.stories): migrate to tailwind
